+++ src/hera/image.py
@@ -1,7 +1,7 @@
 from enum import Enum
 
 
-class ImagePullPolicy(Enum):
+class ImagePullPolicy(str, Enum):
     """A representations of the pull policy for a container and the tag of the image affect.
 
     This influences K8S's behavior when the kubelet attempts to pull (download) the specified image.